created folder pages, created folder comps, moved files, created file Page1.jsx, created file Page1.css
