+++ src/pages/Page1.css
@@ -1,0 +1,7 @@
+#background{
+    height: 100vh;
+    width: 100%;
+    margin: 0;
+    padding: 0;
+    background: rgb(128,128,128);
+}
